chore: rename apps/terminal to apps/console
Reflects the app's identity as a SSH ops cockpit (agent health, deploys,
billing, alerts) distinct from Studio (desktop editor/workshop). Mechanical
rename across Go module path, imports, Dockerfile, CI workflow, release
workflow, and docs.
