+++ README.md
@@ -3,15 +3,15 @@
 Native developer tools for [RevealUI](https://github.com/RevealUIStudio/revealui). One product, two interfaces.
 
 **Studio** — Desktop AI editor and agent coordination dashboard (Tauri 2 + React 19)
-**Terminal** — TUI client for API integration, agent management, and deployment (Go + Bubble Tea)
+**Console** — SSH TUI ops cockpit: agent health, deploys, billing, alerts (Go + Bubble Tea)
 
 ## Architecture
 
-The harness daemon coordinates AI agents, manages PTY sessions, and routes tools. Studio and Terminal are different UIs for the same daemon.
+The harness daemon coordinates AI agents, manages PTY sessions, and routes tools. Studio and Console are different UIs for the same daemon.
 
 ```
 ┌─────────┐     ┌──────────┐
-│  Studio  │     │ Terminal  │
+│  Studio  │     │ Console  │
 │ (Tauri)  │     │   (Go)   │
 └────┬─────┘     └────┬─────┘
      │   JSON-RPC     │
@@ -30,5 +30,5 @@ The harness daemon coordinates AI agents, manages PTY sessions, and routes tools
 
 ## License
 
-Studio and Terminal: MIT
+Studio and Console: MIT
 Harness Daemon: FSL-1.1-MIT (Fair Source, converts to MIT after 2 years)
