+++ apps/console/Dockerfile
@@ -5,11 +5,11 @@ COPY go.mod go.sum ./
 RUN go mod download
 
 COPY . .
-RUN CGO_ENABLED=0 GOOS=linux go build -o /terminal .
+RUN CGO_ENABLED=0 GOOS=linux go build -o /console .
 
 FROM alpine:3.21
 RUN apk add --no-cache ca-certificates
-COPY --from=builder /terminal /usr/local/bin/terminal
+COPY --from=builder /console /usr/local/bin/console
 
 EXPOSE 2222
-CMD ["terminal"]
+CMD ["console"]
