+++ apps/console/qr/render.go
