+++ apps/console/qr/render_test.go
