+++ apps/console/tui/model.go
@@ -10,8 +10,8 @@ import (
 	"github.com/charmbracelet/ssh"
 	gossh "golang.org/x/crypto/ssh"
 
-	"github.com/RevealUIStudio/revdev/apps/terminal/api"
-	"github.com/RevealUIStudio/revdev/apps/terminal/qr"
+	"github.com/RevealUIStudio/revdev/apps/console/api"
+	"github.com/RevealUIStudio/revdev/apps/console/qr"
 )
 
 // ---------------------------------------------------------------------------
