+++ apps/console/tui/model_test.go
@@ -3,7 +3,7 @@ package tui
 import (
 	"testing"
 
-	"github.com/RevealUIStudio/revdev/apps/terminal/api"
+	"github.com/RevealUIStudio/revdev/apps/console/api"
 )
 
 // ---------------------------------------------------------------------------
